Automated Type Def Creation from SQL and C# & Misc Fiddling
~Side note~
First startup is slow, need to optimize the loading of the SqlTemplates (which pre-load each query) or check that they aren't querying the entire table... lol, and limit to like 20 by default with optional override like some of the new Templates "P:\~Dev\Script-Source\~SQL\Query-Files\l"  /shrug

~Main note~
For JS facing functions to know exactly which data is being returned. Generate functions with type hinting annotation for params and type checking before requesting to execute the sql on the backend.
Fiddling about with various methods, including an sql parser, but that's dependent on the flavor of sql, nor does it know which fields are in a table if "*" is provided so the resulting columns and their values cannot be computed that way.
All this to ensure type safety, which makes things easier later.
The system for defining an Sql query method hasn't been setup yet. Will probably involve generating .cs files like the js, just so there isn't duplication of code function everywhere. Ultimately, one thing get defined in a modifiable text file (external of the program) which then creates everything from that (no changing code).

The foundation and structure is taking a lotta thinking and trying things, but should end up way better than the failed system for handling things in OrderManager which just resulted in spaghetti and confusion.

UI has not been touched, but when worked on will have a system for binding data source values to certain elements and communicating to the backend to save those changes. That will be the backbone of all of the UI, and difficult.
Though DataTables for viewing all the item lines and such will be the polar opposite.