Migrated to WPF Project Type (from WinForms) in Pursuit of Using Native UI Instead of Web- More Direct Communication of Data Between Back & Front End
~ weeee ~
All web elements removed, not much investment was made into yet anyways.

WPF Test Project was the basis of dipping the toes and then building out the existing UI, which was then merged into the existing Project/repo- replacing the sln and csproj files as those dictate the project type (which had changed).
The WPF front-end frame work will be quicker to build out UI and also allows for directly interfacing with the data objects, being bound 1-to-1 with them, so less work needs to be done to facilitate that. Not to mention more flexibility and features with the designing the visuals and rendering engine is more updated so effects and such are more performant and built-in.

The abilities of the program currently are being able to search a job, and have it list out all of the Item Lines from the database. Today was manly a fiddling with styling, as I didn't wanna get too deep before I organized the styling files (in the Resources folder) to keep the main files clean.

Next will be the fetching of Order Info to populate the header (which is just static placeholder data). I'll need a way to convert from the GUIDs for materials/finishes to their codes (to display) then bad to their respective GUIDs (which is what the db stores for materials) when changed and saved. Then a save function after the header can get populated.
The ability to push the updated Item Lines (and only what has changed) will likely be after or before the header save function.
The lookup feature (relating to Item numbers, materials, etc) will be a pretty big thing, compared to each other component. Will need to decide on a popup window of sorts, similar to how old crm works now, to filter the respective type of thing and change the corresponding information. But that'll likely be close to last, depending on is other systems would be reliant on it existing first.

-