Figuring out my own mistake which cost me hours to figure out, but Header Data Binding is now setup to work.
Made a critical error which caused the program not to run, making me think that something unrelated was to blame.. it was indeed not lol.
Modified the QueryBuilder CodeGen to output and post process (after building the code files) which utilizes an analyzer which generates the PropertyChanged events + boiler plate required for the DataBindings to correctly be linked update TwoWay.
The ItemLines DataGrid was easy to setup b/c that Element provides an ItemSource property which interfaces directly with a list of objects and isn't normal DataBinding (which the other elements need to use for a single value).
Future DataBinding objects will be quite easy to setup now that I suffered and got it down.
Changing and saving information (and item lines) is ones of the next few things, but I think getting all DataBindings setup first is wise and makes sense.