Eng UI DataBindings Complete & EngOrder UI moved to Page Element & OrderSearch Page Started & Eng Item Lines Filters Implemented
DataBindings for the UI of the EngOrder Page has been linked up. Most of the dropdowns for finishes  has been linked up to a static set of finishes for their type.

All Existing UI moved out of MainWindow and into a Page called EngOrder that gets referenced via a Frame in MainWindow. Modularizing the content and Isolate the behavior and events of a particular Page into itself to keep it all contained and less cluttered.

Page for OrderSearch has been created but is empty and unused currently. It'll work like Old Crm, in that: you search a job, get a resulting list, and click one to open up.

Item Line filters have been setup in a basic capacity. When Ctrl+F is pressed (when focusing the item lines data grid), a bar will popup above and focus a textbox to type in. The textbox will filter upon pressing the Enter key. Filters can be done by just typing a single word, or multiple words with a "+" between- this will be treated as a group and all things must match somewhere in the item line. Multiple filters can be typed into the box by having a space between them (which means you cannot filter a sequence of words with spaces). When there are multiple filters, it will try to match any of them, and will be shown even if only one group matches. Item lines that do not match the filters are hidden from the data grid. To remove the filter, clear the textbox and press enter- the X button only hides the bar and doesn't clear the filter. I added a button and extra space to have some presets, and the possibility to save your own. But I doubt that'd be really be needed as there are only so many you'd need/want to save, so it'll start with presets such as: Worksurfaces, Panels, Assembled, Shiploose, etc- filters that the CRM Copy Line program has currently.