OrderSearch Filtering Recent Orders with JobNbr, Managing the UI Working as intended with the Themes, Troubleshooting TabControl Element
Lot of back and forth with new additions getting removed for them to be altered.

Methods for sending updated data to database is next on the list, as well as looking into a Page for the ProductCatalog as that'll still need to be accessed by Eng (adding items to explode from products and such).

TabControl element makes switching tabs (yes, tabs were added) slow as it unloads the previous content from memory when switching. So going to replace that with some buttons and hiding/showing the respective content, but need to still do that.