Item Lookup Popup & Page Switching Perf
Lookup
- Implemented for material guid lookups in header.
- Not yet implemented for Item line ItemNbr lookup.

Perf:
Added multiple Content Control Elements for each page (not technically an ideal way to do it), since the EngOrder page is so chonk it takes an actual near second to unload when deselected and another when reselected. Switching is near immediate, with the trade off being marginal ram usage increase.

Misc:
- Header data changes not yet save-able, even though the lookup values are modifiable.