Framework for sending network msgs between program instances
Will be useful to either send statuses, updates, or other information about actions that the program performs on behalf of the user. Or to notify someone of something, force-close [with warning] the application to update, check if a job folder is on the client's C drive- just some random examples.
Current implementation is more like a chat, but that was just to make sure sending was working right.