General Clean Up, Consistency, "Auto" Heights to Static Integers, Fixes, and Classes for future use
Removed most of the "Auto" Height Xaml values and replaced them with static numbers to increase the Layout performance of some of the UI.
For loops changed to foreach's for consistency and speed.
Added a "fistLoad" bool value to some Pages to have a reliable way of knowing if something is loading for the first time or not.
Classes for Groups, Users, Permissions, AccessLevel, and Views were added for future. Will allow for Group or User specific access to forms altogether or ability to change certain data within a form.
Paylocity Link added to shortcuts on Home Page
EngOrder, IDNbr column corrected to Idnbr (to match new db).