Unused Code Cleanup & Product Catalog Page Progress
Quick-fire notes:

IcProductCatalog DTO class needed to have the same mumbo-jumbo added to it that the ColorSetInfo class did, so that property changes fire events that can be captured to know what to save. (if those classes are re-generated from the Database for whatever reason, this step will need to be redone, but it's relatively quick even though it's manual)
Page loads the Product Bom and Mfg Items information in two DataGrids which are togglable
Header is prepped to add UI elements within and bind them to the ProductInfo's properties.
DataGrid is collapsible so that all the Header information can be shown in full.
Some Misc styles moved to App.xaml so they could be used by Product Catalog Details page.
Product Catalog Search shows Type and SubType but they don't have dedicated inputs to filter/search by but are sortable.